# thesishub-application/api-docs/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.1-SNAPSHOT",
        "title":"Thesis Hub Application"
    },
    "host":"localhost:8080",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/applications":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/TopicApplication"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/TopicApplication"
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "headers":{
                            "Location":{
                                "type":"string"
                            }
                        },
                        "schema":{
                            "$ref":"#/definitions/TopicApplication"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "500":{
                        "description":"Internal Server Error",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/applications/{id}":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/TopicApplication"
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/TopicApplication"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/TopicApplication"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "204":{
                        "description":"No Content",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/applications/{id}/tasks":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Task"
                            }
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
...
        "/topics/{id}":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "type":"object"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/Topic"
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/Topic"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/Topic"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    },
                    "406":{
                        "description":"Not Acceptable",
                        "headers":{
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "204":{
                        "description":"No Content",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/topics/{id}/applications":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/TopicApplication"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
        "Company":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "plan":{
                    "$ref":"#/definitions/CompanyPlan"
                },
                "size":{
                    "type":"string",
                    "enum":[
                        "CORPORATE",
                        "MEDIUM",
                        "SMALL",
                        "STARTUP"
                    ]
                },
                "url":{
                    "type":"string"
                }
            }
        },
        "CompanyPlan":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "maxTopics":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "Country":{
            "properties":{
                "name":{
                    "type":"string"
                },
                "tag":{
                    "type":"string"
                }
            }
        },
        "Faculty":{
            "properties":{
                "id":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                },
                "university":{
                    "$ref":"#/definitions/University"
                }
            }
        },
...
        "Task":{
            "properties":{
                "application":{
                    "$ref":"#/definitions/TopicApplication"
                },
                "completed":{
                    "type":"boolean"
                },
                "deadline":{
                    "type":"object"
                },
                "id":{
                    "type":"integer"
                },
                "title":{
                    "type":"string"
                }
            }
        },
        "Topic":{
            "properties":{
                "academicSupervisors":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/User"
                    }
                },
                "creator":{
                    "$ref":"#/definitions/User"
                },
                "degrees":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/TopicDegree"
                    }
                },
                "description":{
                    "type":"string"
                },
                "enabled":{
                    "type":"boolean"
                },
                "id":{
                    "type":"integer"
                },
                "secondaryDescription":{
                    "type":"string"
                },
                "secondaryTitle":{
                    "type":"string"
                },
                "shortAbstract":{
                    "type":"string"
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "title":{
                    "type":"string"
                }
            }
        },
        "TopicApplication":{
            "properties":{
                "academicSupervisor":{
                    "$ref":"#/definitions/User"
                },
                "degree":{
                    "$ref":"#/definitions/TopicDegree"
                },
                "faculty":{
                    "$ref":"#/definitions/Faculty"
                },
                "grade":{
                    "type":"string",
                    "enum":[
                        "A",
                        "B",
                        "C",
                        "D",
                        "E",
                        "F",
                        "NOT_FINISHED"
                    ]
                },
                "id":{
                    "type":"integer"
                },
                "link":{
                    "type":"string"
                },
                "officialAssignment":{
                    "type":"string"
                },
                "student":{
                    "$ref":"#/definitions/User"
                },
                "techLeader":{
                    "$ref":"#/definitions/User"
                },
                "thesisFinish":{
                    "type":"object"
                },
                "thesisStart":{
                    "type":"object"
                },
                "topic":{
                    "$ref":"#/definitions/Topic"
                }
            }
        },
        "TopicDegree":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "University":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "url":{
                    "type":"string"
                }
            }
        },
...
        "User":{
            "properties":{
                "company":{
                    "$ref":"#/definitions/Company"
                },
                "email":{
                    "type":"string"
                },
                "faculty":{
                    "$ref":"#/definitions/Faculty"
                },
                "id":{
                    "type":"integer"
                },
                "lastLogin":{
                    "type":"object"
                },
                "name":{
                    "type":"string"
                },
                "phone":{
                    "type":"string"
                },
                "roles":{
                    "type":"string",
                    "enum":[
                        "AC_SUPERVISOR",
                        "ADMIN",
                        "COMPANY_REP",
                        "PROJECT_LEADER",
                        "STUDENT",
                        "TECH_LEADER",
                        "UNIVERSITY_AMB"
                    ]
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "username":{
                    "type":"string"
                }
            }
        }
    }
}